Add themes and plugins auto-updates in 5.5
Just wanted to complete the current sentence which is only mentioning core auto-updates.
For the record the general project of core, plugins and themes auto-updates started in 5.5 and was achieved in 5.6.

This project was conducted via a feature plugin I created: WP Auto-updates (https://github.com/WordPress/wp-autoupdates). It was one of the 4 main goals for 2020 identified by Matt, and I focused most of my 2020 Core work on this, so it would be nice to not forget about plugins and themes auto-updates :) thanks!